fix(aurora-portal): fix ceph tests
